Change how HTTP requests are hijacked in test mode
Instead of configuring octokit.Client with the local test server
endpoint, configure it with the regular GitHub API endpoint but also
institute a custom http.Transport that will redirect all requests to the
local server while keeping the "Host" header and the original HTTP
scheme in the "X-Original-Scheme" header.

This ensures that no test requests can break through into the Internet.